Tiny refactor.
