+++ bin/starter.dart
@@ -1,8 +1,9 @@
 import 'package:args/args.dart';
 import 'package:kernel/ast.dart';
 import 'package:kernel/kernel.dart';
-import '../lib/transformer_wrapper.dart';
-import '../util/dill_ops.dart';
+
+import 'package:aspectd/src/transformer_wrapper.dart';
+import 'package:aspectd/src/util/dill_ops.dart';
 
 const String _kOptionInput = 'input';
 const String _kOptionOutput = 'output';
